Move function under adapters folder
