+++ apps/functions-subscription/src/main.ts
@@ -1,5 +1,5 @@
 import { app } from '@azure/functions';
-import { InfoFunction } from './functions/info';
+import { InfoFunction } from './adapters/azure/functions/info';
 
 // eslint-disable-next-line functional/no-expression-statements
 app.http('info', {
